+++ CHANGELOG.md
@@ -10,6 +10,7 @@ UNRELEASED
 
 ### Changed
 
+* All memory usage in the system is taken into account, not only that by Mesos executors.
 * Only report first load threshold detected to be exceeded, not all.